+++ src/main/java/com/utn/springboot/billeteravirtual/exception/SaldoInsuficienteException.java
@@ -1,14 +1,16 @@
 package com.utn.springboot.billeteravirtual.exception;
 
-import com.utn.springboot.billeteravirtual.model.cuentas.TipoMoneda;
+import com.utn.springboot.billeteravirtual.types.TipoMoneda;
+
+import java.math.BigDecimal;
 
 public class SaldoInsuficienteException extends RuntimeException {
 
     private final Long idCuenta;
-    private final Double saldo;
+    private final BigDecimal saldo;
     private final TipoMoneda tipoMoneda;
 
-    public SaldoInsuficienteException(Long idCuenta, Double saldo, TipoMoneda tipoMoneda) {
+    public SaldoInsuficienteException(Long idCuenta, BigDecimal saldo, TipoMoneda tipoMoneda) {
         this.idCuenta = idCuenta;
         this.saldo = saldo;
         this.tipoMoneda = tipoMoneda;
@@ -18,7 +20,7 @@ public class SaldoInsuficienteException extends RuntimeException {
         return idCuenta;
     }
 
-    public Double getSaldo() {
+    public BigDecimal getSaldo() {
         return saldo;
     }
